backend/apps/content/serializers.py
from rest_framework import serializers

from .models import FAQ, Page


class PageSerializer(serializers.ModelSerializer):
    class Meta:
        model = Page
        fields = ["id", "slug", "title", "content", "meta_description", "is_published", "updated_at"]


class FAQSerializer(serializers.ModelSerializer):
    class Meta:
        model = FAQ
        fields = ["id", "question", "answer", "category", "order", "is_published"]
